backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
]
